requirements.txt
alembic==1.13.1
aws_secretsmanager_caching==1.1.3
bcrypt==4.0.1
click==8.1.7
fastapi-cache2[redis]==0.2.2
fastapi==0.114.1
passlib==1.7.4
psycopg2==2.9.9
pydantic-settings==2.5.2
pydantic[email]==2.9.1
PyJWT[crypto]==2.9.0
sqlmodel==0.0.22
uvicorn==0.30.6
